moving column lookups to dataset builder.
