+++ api/src/test/java/org/pmiops/workbench/api/DataSetControllerTest.java
@@ -93,6 +93,7 @@ import org.pmiops.workbench.model.DataSetExportRequest;
 import org.pmiops.workbench.model.DataSetPreviewValueList;
 import org.pmiops.workbench.model.DataSetRequest;
 import org.pmiops.workbench.model.Domain;
+import org.pmiops.workbench.model.DomainValue;
 import org.pmiops.workbench.model.DomainValuePair;
 import org.pmiops.workbench.model.EmailVerificationStatus;
 import org.pmiops.workbench.model.KernelTypeEnum;
@@ -1011,5 +1012,25 @@ public class DataSetControllerTest {
     final Map<String, QueryJobConfiguration> result =
         dataSetService.generateQueryJobConfigurationsByDomainName(dataSetRequest);
     assertThat(result).isNotEmpty();
+  }
+
+  @Test
+  public void testGetValuesFromDomain() {
+    when(bigQueryService.getTableFieldsFromDomain(Domain.CONDITION))
+        .thenReturn(
+            FieldList.of(
+                Field.of("FIELD_ONE", LegacySQLTypeName.STRING),
+                Field.of("FIELD_TWO", LegacySQLTypeName.STRING)));
+    List<DomainValue> domainValues =
+        dataSetController
+            .getValuesFromDomain(
+                workspace.getNamespace(), WORKSPACE_NAME, Domain.CONDITION.toString())
+            .getBody()
+            .getItems();
+    verify(bigQueryService).getTableFieldsFromDomain(Domain.CONDITION);
+
+    assertThat(domainValues)
+        .containsExactly(
+            new DomainValue().value("FIELD_ONE"), new DomainValue().value("FIELD_TWO"));
   }
 }
