+++ ui/src/app/pages/data/data-set/dataset-page.tsx
@@ -507,7 +507,7 @@ const DataSetPage = fp.flow(withUserProfile(), withCurrentWorkspace(), withUrlPa
           ({domain: domain, values: valueSet, survey: survey}),
         domains,
         await Promise.all(domains.map((domain) =>
-          conceptsApi().getValuesFromDomain(namespace, id, domain.toString()))));
+          dataSetApi().getValuesFromDomain(namespace, id, domain.toString()))));
       return valueSets;
     }
